Layout - Fix placement of selection circles around dots
Previously could not properly place the selectedCoordCircle elements
to share a center with the dot elements.  Issue was that the border
used for debugging their actual positioning changed their
positioning in a manner ever so slightly dependant on the current
font-size scaling factor.  Used a background color instead to
debug to figure out the issue solution.