Gameplay - Added move undo functionality and button
The "squares" state was moved up from the GameBoard component up
to the Game component, along with a few functions that worked
exclusively on this state.  This allowed us to create a button
component as a child of Game with access to the state and with
the ability to undo actions, using a new state variable moveStack,
an aptly named stack of previous moves.