State decoupled from presentation and abstracted to utility.js
In anticipation of changes about to be made to the program,
including the ability to face an AI opponent and to be able to play
others, it was necessary to separate Game state from display logic.

While already technically separate, the game state was stored in a
format very difficult to work with outside of being displayed in
React.  A new file was created, utility.js, housing a new "Move"
type and another "SquareGrid" type to hold the board state, along
with a collection of helper methods.

To facilitate the use of this class in the App, we now use the
React Context API to store the game state as a context.  All of this
has simplified the logic in app.js, removed the need for a lot of
functions there, and allowed us to have the game board state in
a format much more conducive towards manipulation by AI agents
and transmission over a connection