Game Logic - Add support for AI players
A lot of code was refactored within the Game component of the
application in order to allow it to play nice with a new
Player class introduced in players.js that allows for players
to not just be those selecting moves on the screen, but
arbitrary move-making agents.

A very basic AI agent that makes random moves has also been added!