Game AI - Changed the AI agent from RandomPlayer to a new WeakAI
Changed the AI agent to a more advanced one that at least won't
be giving anyone any free boxes, and will always take any if you
set them up for it.  A step up from the RandomPlayer AI.

A few bug fixes in code in utility.js that hasn't been used much
until now by this new AI agent has also been changed.  Webpack
now produces source maps, which has aided greatly in debugging,
but these source maps are not fit for production, so this should
be changed to not happen in production as soon as possible