Refactor - IN PROGRESS - Major change in App.js structure
There's about to be a good number of changes coming up soon,
namely the introduction of a new component meant to control
the Game object and take over some responsibilities for it,
a new menu system for the selection of game options, and
support for remote mulitplayer.

In order to support these changes and simplify code, the main game
code got moved out of App.js and into its own GameComponent.js
Some things that were previously internal to the Game are now going
to be externally controlled, and GameComponent.js (previously
mostly App.js) has been updated and provided with TODOs indicating
the locations likely to be changed soon.

[ CHANGES IN PROGRESS ]