WIP Backend: Change GameComponent.js to use funcitonal components
Major refactor in progress here.  The desire is to move from a
traditional react state model augmented with context to one where
the context is accessed through the useReducer() hook, and the
state management logic is refactored into the reducer code,
leaving the components, especially the topmost one, much easier
to understand, clarifying the code, and reducing my dependence
on callbacks, instead being able to use dispatch and the global
Context data store when it makes sense, leaving state that's meant
to be local internal to the components but extracting state that is
not.

Currently have written the reducer and context objects, and just
finished converting the top most Game react component in
GameComponent.js into a functional component, and moved the previous
internal methods that handled state changes into the reducer,
allowing me to create much simpler and easier to understand code,
also code that'll be easier to extend once I add planned features,
like online play.

In fact, it is the initial attempt to add online play that fueled
my desire to refactor, because the state management logic and the
logic for communication between the players and the Game component
was very convoluted layers of callbacks and stuff, now it's just a
dispatch function that's sent over, and a few added context vars.

-Andy