WIP Backend: Change GameComponent.js to use functional components
See last commit.  What's changed since is that GameComponent.js
has been fully converted to functional and hook-using form.
What remains to be done is to update the player objects to use
nextTurn() and endTurn(), and to establish a way for the Game
component to subscribe to Player move updates.  (Not _that_ hard
to figure out, but right now brain is dead so leaving it for next
time)

-Andy