Change GameComponent.js to use functional components
WIP Backend: Change GameComponent.js to use functional components II

Finally, all of the GameComponents have been changed to functional
components.  Almost all of the game logic has been removed from
these functional components and instead shifted into GameContext.js
This gives us the benefit of having much cleaner and easy to read
code, and also the added benefit of easier debugging (since we have
essentially just created a state machine, and those are easy to
work with), and also much easier testing (we can simply mock a state
and send in commands to the reducer and see that the final state is
as it should be).