Feature - Add ControlPanel feature to main branch
It has for a very long time been a desired feature for there to be
some sort of side panel, besides the main Game/GameBoard panel in
GameComponent from which to keep track of score and have controls
for the game / ability to start a new game from, and it is now
a reality.

Along with this there also is a GameMenu, which is both more
streamlined and capable, and also more extensible with CSS and
additional options, compared to the old method using javascript
prompts.  This update sets the game up for the introduction
of many new features, such as online play setup.