Refactor - Change OwnershipGrid into TaggedGrid
A new class has been added into utility.js in order to replace the
old OwnershipGrid.  The idea is that we're going to want to
potentially have different types of tags to attach to a given
coordinate on the game board in the future _besides_ initials
indicating ownership of a square, and instead of introducing a new
data structure for each one, we can simply have a "tags" data
structure, where each coordinate has an associated dictionary
of tags that can be accessed, one of which are initials
corresponding to a player that has taken the location.

Previous code in GameComponent & GameContext has been changed to
use this new class instead.