Reducer - Move startTurn() from reducer to useEffect
While working on a feature branch, it immediately became evident
that calling startTurn() on a Player object from the reducer
in GameContext.js was a bad idea.  Reducers shouldn't have side-
effects, but mine does.  The other side effects are rather
innocous, though.  This one is not, because the startTurn()
method triggers many things, some of which in the future may
themselves want to launch dispatches; dispatching from within
dispatch is a big no-no, which is why the __runBatchedActions
action exists in the first place.

To remedy this problem in the least-disruptive way possible,
a new state variable was created, numberMovesCompleted for the
sole purpose of a useEffect() in ControlPanel to watch and then
itself call startTurn() on the correct Player.  This way, we
no longer have any restrictions on what startTurn() can do,
allowing it to issue its own dispatches if it pleases [ although
it cannot dispatch a state change to numberMovesCompleted, lest
we be stuck in a loop -- this restriction is obvious though[.