Feature - New "App Settings" feature added
This feature consists of state held in the App component that
is preserved between launches of the React Appliaction, and a
GameMenu component that is displayed upon pressing a "Settings"
button in the ControlPanel.

There are currently two settings, "Debug Mode" and
"Save Previous Match Settings" that are booleans that can be set
and persist between sessions, but whose effects have yet to
be implemented;  they are to be implemented in coming commits.