Game Menus - Add defaults and navigation to menu
GameMenu now has a button allowing backwards navigation, buttons
for exiting menus have been removed as they were not functional
and were not intended to be used anyway.  Also, GameStartMenu
now has an option to shortcut to saved/default settings on
the first screen, to make rematches easier to start.