Bugfix - Fixed settings bug introduced by previous commit
The previousSettings commit introduced a bug where if you're
saving and loading your settings from localStorage, and you
previously played a match with a CPU opponent, you wouldn't
be able to switch back to playing a regular opponent.  The reason
for this was that the original menu logic was fragile and didn't
account for the possibility of gameType being set to "CPU" by
default, the logig always assumed it was unset.  Now, it is
explicitly set to either "local" or "CPU", as it honestly
should have always been for the sake of clarity.