AppSetting - Implement & Create "Show Player Range"
A new AppSetting has been created that saves additional information
within the moveHistory state object in the GameContext: a range of
moves that a Player agent was considering to do.  This doesn't add
additional information for LocalHumanPlayer(s), but for the AI Player
agents, if they picked from a random set of possible moves to do, it
shows the range from which that move was selected.