Testing - Add AI Player move decision testing code
A new series of tests for AI players has been added to players.test.js
These allow you to feed in a moveHistory, and ensure that the move
ranges produced by the AI player being tested match the expectations.

Included is one example test on a simple MoveHistory given in the
test fixtures to show how this test functions.

This test _is_ currently running slow, however (~3000ms at the time
of writing this commit message), some changes to the code might
be necessary if we want to be able to run several tests -- at
first glance, since we have the entire move history, every step
in the move history can be tested simultaneously, for example.