TESTING - Improved AI Player testing routines
BasicAI testing code has been refactored, its logic is now much
much easier to follow and it also allows for all of the AI turns
in a given MoveHistory to be tested simultaneously and independently

Also, a small typo introduced in src/players.js after commit editing
has been fixed.  This is part of the reason to push for more testing:
by testing more parts of the app automatically, less mistakes will
make it onto master.