Refactor - Change SquareGrid to no longer save moveHistory
Since moveHistory is now its own object and also keeps track of
player numbers for each turn, we no longer need the built-in
moveHistory feature of SquareGrid.  It never saw much use anyway.

It still is sort of being used in one series of tests;  these
changes do not change those tests, but in a future commit they
will be ported over to the new system of seperate SquareGrid
game states and MoveHistory objects.