Players : Fix BasicAI bugs
Three bugs in BasicAI were fixed; one was an issue in the helper function
groupMovesIntoTaggedChains(), which while not formally part of the
BasicAI class is essential to its functioning.  It was incorrectly
grouping moves that completed squares into the current chain even if
they weren't connected.

Another bug was the strategy "freeDoublyActive3ChainMoves" -- no such
category of moves exist; one can choose to not take these as to then
take a longer chain.  Hence, it was removed from strategies but the
function kept in players.js just incase there is in the future a
reason to re-instate it.

The last bug (technically two bugs) were in the two BasicAI strategies
that calculate a quantity known as the "Control Value" -- potentially
because previously mentioned bugs screwed up the functioning of BasicAI,
these were left to slide.  The previous branching on CV being greater
than a certain number was actually the opposite of the logical/desired
behavior.

A test case was added, and a previous test case changed in order to
reflect the changes in BasicAI behavior.