Refactor - GameBoard responsibility clarification
There are a whole bunch of little changes that are difficult to
disentangle and hence all are in this one commit, but the underlying
theme of this commit is to more clearly deliniate the responisibilities
of different parts of this application.

In short, ControlPanel (and ControlPanel.js) were starting to get very
bloated from an excess of responsibilities.  It was was a convenient
place to dump things because it's responsibile for the menus and setting
the score and, but this was making the code very confusing and harder to
debug.

This app has, roughly, three components:
1.  Main game logic
2.  Game Presentation logic / input logic
3.  AI player logic

The issue was that a lot of #1 was bleeding into components meant
for #2.  This commit is the first in a series meant to remedy this;
every useEffect was moved out of ControlPanel except for the one that
loads settings, since that one is directly relevant to this component,
which contains the current settings state and is reponsibile for setting
up menus that let us alter state.

These useEffects were moved into the Game component, but only
temporarily; soon a custom hook that will reside in GameContext (so that
the game logic can truly reside in one place) will replace the
useReducer there, and will contain all of these useEffects within it.

Also, since the Game component is now the top-most component for game
logic, it was split away from the rest of the components, which instead
decribe the game board state and user interaction with it, which was
refactored into a different file.  Also, the Different menus were
factored out of ControlPanel and into their own files as well.  This
way, each file is a coherent idea, making it easier to change the app
and read the code.