Refactor - State side-effects moved into custom hook
Previously, side effects of reducer action that could not be done
inside of the reducer itself were moved off to useEffect(s) that were
scattered about in either the Game component or the ControlPanel
component (usually the latter).

The previous commit moved them all out to the Game component, and this
commit moved them all into a custom useGameStateStore hook that resides
in GameContext.js .  This finally achieves the goal of completely
separating the core game logic from the presentation/input logic, as
desired.

Also, after the removal of the last bit of core game logic from the Game
component, it became evident that the component itself had become
redundant, serving the same purpose as its containing App component;
hence it has been merged into the top-most App component.