WIP - New GameMenu code
This is a very-much WIP commit of a overhaul of the GameMenu system; in
it we navigate away from using bootstrap-react-carousel and
bootstrap-react-modal to a custom-made modal and menu transition
animation instead done using react-transition-group (itself a dependency
of boostrap-react-carousel) and CSS transitions.

It basically works in its current form, but there are CSS changes to be
made; also, since the menu panels aren't in the DOM until they're needed
(This may be changed in the future), GameStartPanelMenu's populating of
the selects has been broken and needs to be fixed.