GameMenu - Complete implementation overhaul
GameMenu has been greatly simplified by ditching the previous Carousel
and Modal elements borrowed from React-Bootstrap, instead providing the
desired menu item transition animations through the use of CSS and
react-transition-group components SwitchTransition and CSSTransition.

Also, to simplify the code even further we've changed how GameMenu works
a slight bit -- instead of taking in a render prop that returns a
fragment, it now insists on an array of elements instead; this allows
GameMenu to have to do less work on each render.

On top of these changes, there have also been a few changes to
GameStartPanelMenu, namely the first two GameMenuItems contain buttons
styled as links isntead of actual links and the select HTML elements are
fully generated from the beginning and passed into GameMenu fully formed
instead of waiting for them to be generated in the DOM to be then
properly populated ; with the change in GameMenu the former method is no
longer viable, not to mention it's the less elegant / more complicated
way of going about it.