WIP - SelectableTable functionality check
It's ugly as sin, but the SelectableTable is indeed working as a
component as intended!  There's still a ton I'd like to change about the
layout, css, etc... but the core functionality is definitely there.  I
really can't wait until I have time to really work on the CSS for this
thing because it could be so much less ugly, lmao.