GameMenu -  Change GameMenu items prop format
The GameMenu items prop format has been modified -- instead of taking an
array of GameMenuItem(s), it now instead takes a render prop returning
an object, with the key names being the name of each GameMenuItem, and
the value being the items themselves.

This greatly simplifies code in GameMenu, letting us completely
eliminate the previous conversion to integer page numbers.  This was
honestly a relic of a previous design that used Carousel objects in
order to enable moving between GameMenu pages.  It makes no sense to
refer to them by integers, since all of the link explicitly name which
page they move to; this is much easier to follow/understand.