Online - Fix bug with not sending some moves
In the interest of being overly cautious and avoiding a bug that
hadn't manifested yet, RemotePlayers would not send a performed
move done by the LocalPlayer to the server until it was the other
player's turn, saving it instead in a local "lastMove" class var.

The result of this, however, was that when a player performed
multiple moves in a row (which is legal when capturing boxes),
only the last move would be broadcast, leading to a desync in game
states between the two players.

Since the server does not run an instance of the game, we can't
prevent malicious attempts at modifying the code as it runs in
the browser to purposefully cause desyncs, but they shouldn't
naturally occur in the code!