GameStartPanelMenu - Fix online name chooser bug
In ChooseOnlineNameComponent.js, if you chose a name that wasn't
available, the custom validator was provided with a message letting
the user know that the name isn't available.

However, the value set by the input element's setCustomValidity
method is never cleared in the case of a usable name being selected;
thus, if you selected an unavailable name then tried with an available
one, upon trying to submit the form it wouldn't let you, saying that
the previous name you selected is not available.

This has now been fixed as to clear the validity message by replacing
it with an empty string.