GameContext - Refactor from context to redux-like Store
This is the first of a series of many commits refactoring the code
from an approach based on the use of React Context for holding
global app state to one using a customized redux-like store of
central game state.

There are many more changes in store, but this one simply moves
away from global application state being stored in a top-most
context to it being stored in a redux-like central store.

React Context is meant for use when you have a situation where you
want to render different subtrees of components with different
contexes.  That is not the case with my application, at least in
the circumstances that we were using it in.  I was just wrapping the
entire app in one Context to be available to all, which is unwise.

That, combined with our use of useReducer made for a very inflexible
framework that led me to create unclear code in order to force what
the app was/is trying to do into a box that didn't fit it very well.

Now, instead, there is a custom redux-like store implementation,
coupled with a custom useStore hook.  This store will be fitted with
useful extensions/middleware such as a version of Redux thunks and
the ability to handle arrays of actions [such that state is changed
and broadcast only at the end of the sequence of actions] natively,
as well as allow for subscriptions from non-component objects.

This will greatly simplify many areas of the code and allow us to
dispense of very tortured workarounds which, in addition to being
unclear, unnecessarily coupled parts of the application that didn't
make sense to be coupled and hindered our ability to add features,
as evidenced by the long development time taken up by the online
play feature.