Players - Player objects no longer manage score
It was a huge pain having player objects have to be recreated (since
they are state, and state is supposed to be immutable under react)
every single time that the score changes.

If you think about it, the score isn't a property of the player
object, it is a property of the current game state, which is this
big global store of data.  Players shouldn't have to hold on to
this variable; it can be convenient sometimes for them to, but
it clearly is not in this case.