Player and Data Flow Overhaul
It would have been nice to commit all of these changes separately,
but unfortunately they involve a lot of systems that were linked
together and very difficult to update separately, possibly leading
to an even more confusing series of commit than this one large one.

In essence, this commit turns Player objects from classes into a
plain data object, eliminating all of the oddities they used to
have with state.

It also creates a redux-like central store of data, to which is
stored a new "gameEngine.js" which acts like the necessary useEffect
hooks that managed things like turn advancement.

There's also been the introduction of a new appState state variable,
a more nuanced replacement for isActive.

A lot of code in the reducer was able to be eliminated since it was
no longer needed to awkwardly handle terribly designed player
objects.  The logic handling online connections and turn advancement
and asynchronous ( remote player moves and UI move attempts) moves
and synchronous ( CPU player ) moves has been greatly simplified
as a result of the overhaul, with many more simplifications to come.

The CPU agent functionality which was most of the file players.js
has been moved into a much more aptly named CPU.js file.

In short, everything makes 10x more sense now than it used to, and
the program flow is much much less convoluted and it has become
more modular than before, and the ground has been set for even
further simplifactions and refactoring down the road with the
simpler central data store, simpler player objects, and the
addition of an appState variable that can be used to control the
game through dispatches to the central data store, avoiding the
use of complicated workarounds or tons of useEffect statements with
no real logical home in order to trigger the right actions on
particular state changes.