Control Panel - Removal of redundant Player object code
In the past, because of the unusual situation caused by the awkward
Player objects through which all move attempts had to go, there
had to be an applySettings function that registered callbacks with
the player objects, passing them a copy of its own access to a
central GameContext that let them perform moves.

Now, though, all of the set-up can take place inside of the reducer
since this is no longer necessary since both moves no longer have
to go through Player objects and since you no longer need a reference
to a component in order to alter state.