appSettings - Migration away from App.js and context to central store
The previous large commit had begun to fold appSettings into state
in order to allow reducer actions to act differently depending on
the value of appSettings without having to explicitly have it be
passed in.  This commit finishes this process, removing any
responsibility App.js has with handling appSettings and moving it all
into the central store or reducer actions, and handling the initial
load of appSettings into GameEngine.js.

As the case with the last few commits, this has lowered the total
NLOC and increased clarity.