GameBoard - Fix Error on Online Match Start
Fixes bug introduced in 0CED8E2A

For what seems to be no good reason at all [ this bug only happens
when starting an online match ], there will be two instances of
GameBoard when the state is first updated by setUpGame if the
reducer action is ultimately being triggered asynchronously by
a callback from Socket.js instead of being triggered ( also
asynchronously, technically ) by clicking a button on the
GameStartPanelMenu UI that calls it.

Even if gameStateReducer({ type: 'setUpGame, settings }) is called
with an identical settings object, it does not matter, we still get
this bug (if, for example, Socket.js is modified to always start
a CPU match and ignore all data received from server instead of setting
up an online match ).

After looking through many different possiblities for fixing it and
trying various things, this is the only one that works.  The "extra"
GameBoard vanishes immediately, too.  At least if it stuck around
there would be some sense to it.

This bug may be due to the changes with how state is initialized, and
the moving of some things out of useEffect statements and into
GameEngine.js, but a quick inspection of the code still leaves me
completely puzzled as to why this bug would occur.  Hopefully, in the
future, this commit can be reverted and the code will function without
this fix; poorly understood fixes are bad fixes, but of all possible
poorly understood fixes, this one is the least likely to cause
problems.