Socket.js - Refactor Game Lobby management into new object
Game lobby funcionality and state has been refactored out of
Socket.js, SocketContext.js, and GameStartPanelMenu.js and moved
into a new GameLobbyHandler.js.  GameLobbyComponent is now just
a visual layer and no longer holds presistent state or application
logic.

As I try to add functionality to the app related to online play and
allowing for multiple matches with the same player with a seamless
transition from one into another, I've come to realize that having
the state and relevant methods relating to all of the actions
related to setting up a match between two players and keeping track
of who they are and relaying moves between them all scattered in
different places makes for very confusing code, and makes it hard to
patch new features into it.

Thus, I've created a new gameLobbyHandler which is... essentially
just that.  It handles the client-side state and functions necessary
in order to keep track of and interact with the server-side notion
of a game lobby, putting it all in once place.  Instead of using
persistent state in the GameStartPanelMenu along with gameState and
some state in Socket.js, now it's all just in this new handler.

All of the methods necessary to work with and manipulate this state
also lie here