Start Menu - Game Lobby change to use GameMenu history stack
Previously, whenever a game lobby became disbanded, you'd be kicked
back to one of two menus, depending on whether you were host or not.
This is fine, but the issue is that then the GameMenu history stack
would get polluted with a GameLobbyComponent representing a Game
Lobby that no longer exists/is valid, hence trying to "Previous"
your way back to the first menu screen was impossible, instead you
would be stuck in an endless loop between the GameLobby, which you
would be immediately kicked out of, and either the active games list
or the game hosting options menu.

The code in GameLobbyComponent has been changed to now instead just
use linkTo's ability to send you back to the last menu screen you
came from instead of explicitly linking you to what the last menu
screen you were on should be, thus fixing the bug and simplifying code
in the process.  Double win!