Server - Clear ready status upon match start
The Game Lobby on the server is supposed to remain between matches,
so that players don't have to go back to an active game list to
play again.  We now clear ready status upon match start in order to
fix a bug where when players wanted to play a second match, it
required just one player to be ready instead of both.